refactor: improve api client and service layer structure

+++ src/data/categoryService.js
@@ -1,13 +1,15 @@
 import { apiClient } from './apiClient';
 
 export const categoryService = {
-    getAllCategories: () => apiClient.get('/categories'),
+    //public endpoints
+    getAllCategories: () => apiClient.get('/public/categories'),
 
-    getCategoryById: (id) => apiClient.get(`/categories/${id}`),
+    getCategoryById: (id) => apiClient.get(`/public/categories/${id}`),
 
-    createCategory: (dto) => apiClient.post('/categories', dto),
+    //Only admin endpoints
+    createCategory: (dto) => apiClient.post('/admin/categories', {name}),
 
-    updateCategory: (id, dto) => apiClient.put(`/categories/${id}`, dto),
+    updateCategory: (id, dto) => apiClient.put(`/admin/categories/${id}`, {name}),
 
-    deleteCategory: (id) => apiClient.delete(`/categories/${id}`),
+    deleteCategory: (id) => apiClient.delete(`/admin/categories/${id}`),
 };
